fix(example): send http request to discounted-cartrade
